requirements.txt
absl-py==0.9.0
astor==0.8.1
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
gast==0.3.3
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
importlib-metadata==1.6.0
joblib==0.14.1
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.1.1
more-itertools==8.2.0
numpy==1.16.4
opencv-python==4.2.0.32
packaging==20.3
Pillow==7.2.0
pluggy==0.13.1
protobuf==3.11.3
psutil==5.7.0
py==1.8.1
PyOpenGL==3.1.5
pyparsing==2.4.6
PyQt5==5.14.2
PyQt5-sip==12.7.2
pyqtgraph==0.10.0
pytest==5.4.1
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.22.2
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tqdm==4.45.0
urllib3==1.25.8
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
